MMT_VSCODE/requirements.txt
flask==3.0.3
flask-pymongo==2.3.0
flask-mail==0.10.0
flask-cors==4.0.1
apscheduler==3.10.4
python-dotenv==1.0.1
